Filter "fake" block entities from blockEntities
Current solution of writing to the ChunkData packet accidentally includes "fake" Block Entities to be counted in the blockEntities.size() var write. Causes issues with clients attempting to load chunks that include non-block entities with a custom handler. This solution proposes that these blocks get filtered out prior to the packet data being written, via a stream -> filter -> collection solution. Open to ideas on how to better solve this issue, as it is impossible to get logs of due to it being a client issue.

To replicate the original issue;

1. Create world containing a non-block entity block (ie; a cobweb) in the chunks where the player will spawn
2. Create BlockHandler for that type of block- example attached below

```java
public class CobwebHandler implements BlockHandler {
	private Random random = new Random();

	@OverRide
	public @NotNull NamespaceID getNamespaceId() {
		return NamespaceID.from("minecraft", "custom_cobweb");
	}

	@OverRide
	public boolean isTickable() {
		return true;
	}

	@OverRide
	public void tick(@NotNull Tick tick) { // super cool cobwebs that play the stepSound packet randomly
		if (tick.getInstance().getChunkAt(tick.getBlockPosition()).getViewers().size() == 0) return;
		if (tick.getInstance().getTime() % (50 + random.nextInt(35)) == 0) {
			EffectPacket stepSoundPacket = new EffectPacket(2001, tick.getBlockPosition(), tick.getBlock().stateId(),
					false);
			tick.getInstance().sendGroupedPacket(stepSoundPacket);
		}
	}
}
```

3. Register the handler as so;

```java
MinecraftServer.getBlockManager().registerHandler(NamespaceID.from("minecraft:cobweb"), CobwebHandler::new);
```

4. Connect to server. You should immediately disconnect with a `DecoderException: z: Loading NBT Data` error.

To validate solution, repeat the the above steps, but including the updated ChunkData class. You'll be able to connect normally and the BlockHandler will tick as normal.