Adding a Paginator component with tests
Paginator is a component to store elements in a list and retrieves them into "pages".
A page is a list containing some elements stored in Paginator.

Example:
I create a new Pagniator containing these numbers with a pageSize of 3:
{0, 1, 2, 3, 4, 5, 6}

Calling paginator.get(1) returns me:
{3, 4, 5}

Calling paginator.get(2) returns me:
{6}

I can also get a specific value in a page or set it...