Revert "Improve entity jump"
This partially reverts commit 6fe078c7708c9fc36049a876fbc0f396f09656ed on TheMode's request.
Kept the Navigator#moveTowards return change, as it could help prevent redundant calculations.

Unfortunately also reintroduces entities floating in the air when pathing on non-full blocks - this is an issue with the Hydrazine pathfinder and not related to Minestom (pathfinder calculates with integers even for the Y axis)