Improved socket handling for the server (#1998)
* Improves socket handling from the server.
A logging is added which catches errors as soon as the server socket connection cannot be closed. In addition, the program terminates itself immediately to avoid further errors.

* Worker stop has been transferred.
The worker stop was implemented in a close method in the worker class and is now called from the server. The access modifier was adapted for this

* Adding a wakup call before the server closes its socket