Change how ticking for touch works
Made it so it should only fire once unless the entity is actively attempting to slide along it.

Changed it to use the computed physics for entities.
Use slower brute force calculation for Players as there is no way for the current physics engine to handle XZ without some more work. Possibly a Controller interface that sits on top of Entity?

TODO: Fix Tests as they are currently broken as they compensate for the older system.