Immutable Event system draft.
Sorry for the large commit. It was hard to split up the work as it would either compile or not.

[x] Delete Depercated functions inside events.
[x] Use record name for events unless inside the mutator. This is a heavily breaking change for an already major breaking change.
[x] Allow split mutation, I dont like how this is implemented maybe possible to do something like the network buffers for mutability?

Most of the events need documentation. Any field I found that didnt have any and was getting renamed was deleted.

Seems to work on demo, fails unit tests need more testing. Probably my bad implementation of. Approx 15 tests are failing. (Most due to the updated tester implementation maybe?)

For now this a draft and possibly needs more refactoring. Anyways time for sleep.