Use double lock idiom to prevent racing on cached packets.
Imagine you two players where the viewer is trying to see the same chunk at the same time. Now the chunk has to do twice the calculations and can sheer its data. Even known the implementations are expected to be thread safe, I think it should be up to minestom to ensure that their supplier is only called once if multiple threads are competing.