feat: Faster Socket writing
Locked behind a ServerFlag as it needs more widespread testing outside the demo server. Also, in case there is a regression in peaks
Also exposes the read and write thread of the PlayerSocketConnection for any future use.
Also names all the socket-based virtual threads for the logger.

If testing goes good and no major regressions are found, this feature can be upgraded as the default.