Non-breaking: Add the OptWorldPos metadata type, and misc
Add the OptWorldPos type in case users want to use it (w/mods), as it should be a supported type. This also makes the optional static final inside the WorldPos class as it's used in more than one place now.

Rearranges the Metadata file to put the bytes in another file (Metadatas, yes, I know it's not a word, but it follows the semantics that we have for other files of its types) that can be delegated to autogeneration if we want to remove the nextId)

Also, converts ObjectArray to use a List#of instead of an ObjectArray as its stable; maybe an ObjectArray#toList should be added to return a stable list version?