Squashed Registry Changes
They are kind of blanketing changes like to do immutability on ObjectArray; you need to toList because you cant array copy without refining the class type, which is annoying.

DynamicRegistry can now compress down to something similar to a static registry with unsafe ops disabled. (And not testing of course). This also move the locks into a single monitor, as what are the use cases for individual locks.
Fixes a possible issue with racing on the CachedPacket registry, to be fair; I haven't seen an implementation that needs a revolving registry and cannot just become a StableValue later.

Changes to Map<String, Map<...>> to just be backed by Properties, the keyset is weird, but where do you want to move the casting is the question. Also finally creates getList, and depercated special cases (getNestedDoubleArray). This also changes how Properties are loaded, makes alot of things immutable returners, in general, this is useful for the ItemEntry, as it creates a prototype on the fly; and before this its stays for a long time. It might be worth creating a special holder class for each state in this entry to drop the 96/64 bit header and extraneous data after the first load.

Other changes to entries like making the entity offsets be constants for EntityEntry.

This commit also puts in the defensive copies for known immutable things, only unsure one is the HolderSet implementations, as Holders in general are confusing.

I haven't seen a significant regression in loading times, but that probably should be tested as registries have always been the slowdown; and memory usage for BlockEntry has dropped due to interning key instances.