Use Map instead of PacketRegistry ClassValue
Seemed faster than ClassValue, with 150 bots 3 min test with chat message default in BotMark. 180ms map vs >300ms ClassValue. I am aware that ClassValue is lazy, but im unsure the JVM can optimize it farther because the fixed size is unknown compared to map, and maybe stable function later. I'm going to assume a switch statement would probably be the fastest here.

Also we weren't using the ClassValue main sell of allowing the classes to be unloaded, because we always hold a strong reference regardless in the suppliers now packetsById.