Change Builders to be Settings and become immutable
This should allow users to eventually set a default settings that contains the default arena and default resizing strategy.

This also changes the nullability of registries to throw if you supply extra null arguments (Makes no sense to accept them).

Fixes an issue with how resizing would never free its other memory due to the nature of arenas, currently we only have an arena to just hold a strong reference to it, which seems wasteful in certain contexts.

Add NetworkBuffer#fill if we need to do cryptography which may require filling the end with zeros again.

This also marks the great doc'ning.

Though the Packet test still needs the 200 packets for Play state added, so we can enforce them in the future. I think dialogs are going to be the worst ones here. Maybe we should split them into packages like how the packets are currently layerd out. We still need to generate bin tests for minestom in data generator.

We could also use translation keys in the data generator to put in the kick messages for PlayerSocketConnection.
There is a slight edge case where the players gameprofile; causing NPE during login is missing because the race introduced #2449, but Im unsure how to fix cause its not as simple as putting volatile because its reading before its ever set.