Reoptimize IOUTF8StringType for ascii
Changes name from IOUTF8String to StringIOUTFType which is better.
Readd ascii optimization use built-in impl for reading

While faster for ascii it's going to suffer performance on writing Unicode, but ascii is likely your common usecase, apart from fonts which will be too small where the overhead would be from other places.

Maybe in the future we can introduce a cached type for this expensive operation.