Split resizeable and static implementations.
This makes it so static buffers can never be resized. This also makes it so resizable buffers can never be dummys. This should give the JIT more information, but it may introduce a polymorphism overhead. Ive tried to keep these out of the hot loops.

There shouldnt be too much performance hit since alot of the implementation is left in NetworkBufferImpl which has final methods.

These can never be value classes sadly, since write and read index are tracked by the network buffer.