feat: Unsafe Collections flag for JcTools
Should prevent any warnings from unsafe usage, does NOT automatically enable with JVM flags, as this project shouldn't be reliant on unsafe if available. Waiting for JcTools v5 or etc. that changes to VarHandles.

Native image is in a unique spot with this flag, because either or require different native image traces. The assumption is if your tracing native image either should work, or if your chasing performance you should have this enabled anyways with your native image.